feat(motion_planning):  增加使用矩阵表述包络体以加速检测碰撞检测

+++ src/motion_planning/include/motion_planning/scence.hpp
@@ -15,8 +15,8 @@ virtual bool isCollision() = 0;
 
 struct SphereObstacle:public Obstacble
 {
-double x,y,z;
-double radius;
+float x,y,z;
+float radius;
 bool isCollision()override
 {
 return true;
